Rename Duke to CliRental for class and package name
