+++ build.gradle
@@ -29,11 +29,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("seedu.clirental.CliRental")
 }
 
 shadowJar {
-    archiveBaseName.set("duke")
+    archiveBaseName.set("CliRental")
     archiveClassifier.set("")
 }
