+++ src/test/java/seedu/clirental/CliRental.java
@@ -1,9 +1,9 @@
-package seedu.duke;
+package seedu.clirental;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class CliRentalTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
